+++ FE/src/components/transaction/TransactionAddModal/MessageInput/index.scss
@@ -1,6 +1,7 @@
 .message__input {
   textarea {
     border: 2px solid;
-    min-height: 150px;
+    min-height: 350px;
+    resize: vertical;
   }
 }
