+++ FE/src/components/transaction/TransactionAddModal/TransactionAddModal.scss
@@ -14,8 +14,9 @@
     display: flex;
     flex-direction: column;
     align-items: center;
+    position: relative;
     width: 700px;
-    height: 800px;
+    height: 80%;
     background-color: white;
     z-index: 2;
     border-radius: 30px;
@@ -31,7 +32,7 @@
       box-shadow: 0px 2px 2px 0px grey;
     }
 
-    form {
+    .form {
       display: flex;
       flex-direction: column;
       align-items: flex-start;
@@ -95,6 +96,20 @@
 
     .transaction__cancel__button {
       background-color: rgb(253, 45, 45);
+    }
+
+    .message__input__button,
+    .message__input__back {
+      position: absolute;
+      top: 23px;
+      left: 19px;
+      border: none;
+      background-color: #fff0;
+      color: #0d7ee1;
+      font-weight: 900;
+      font-size: 15px;
+      cursor: pointer;
+      outline: none;
     }
   }
 }
