+++ FE/src/components/transaction/TransactionAddModal/index.tsx
@@ -1,13 +1,7 @@
-import React, { FormEvent, useEffect, useState } from 'react';
+import React, { useState } from 'react';
 
-import { useTransactionData } from '../../../store/AccountBook/accountBookInfoHook';
-import DateInput from './DateInput';
-import CategoryInput from './CategoryInput';
-import PaymentInput from './PaymentInput';
-import PriceInput from './PriceInput';
+import TransactionAddForm from './TransactionAddForm';
 import MessageInput from './MessageInput';
-import ContentInput from './ContentInput';
-import { createTransaction } from '../../../api/transaction';
 
 import './TransactionAddModal.scss';
 
@@ -18,26 +12,7 @@ const TransactionAddModal = ({
   setTransactionAddModal: (value: boolean) => void;
   accountbookId: string;
 }) => {
-  const {
-    categoryPool,
-    paymentPool,
-    accountbook,
-    addTransaction,
-  } = useTransactionData(store => ({
-    accountbook: store.accountBook,
-    categoryPool: store.accountBook.categories,
-    paymentPool: store.accountBook.payments,
-    addTransaction: store.addTransaction,
-  }));
-
-  const today = new Date();
-  const [year, setYear] = useState(today.getFullYear());
-  const [month, setMonth] = useState(today.getMonth());
-  const [day, setDay] = useState(today.getDate());
-  const [category, setCategory] = useState({});
-  const [payment, setPayment] = useState({});
-  const [price, setPrice] = useState(0);
-  const [content, setContent] = useState('');
+  const [message, setMessage] = useState('');
   const [messageInputVisible, setMessageInputVisible] = useState(false);
 
   const closeModal = () => {
@@ -48,77 +23,42 @@ const TransactionAddModal = ({
     if (e.target === e.currentTarget) setTransactionAddModal(false);
   };
 
-  const onContentChange = (e: Event) => {
-    if (e.target) {
-      const targetElement = e.target as HTMLInputElement;
-      setContent(targetElement.value);
-    }
-  };
-
-  const onSubmitClicked = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const date = `${year}-${(month + 1)
-      .toString()
-      .padStart(2, '0')}-${day.toString().padStart(2, '0')}`;
-    const newTransaction = {
-      date,
-      type: '지출',
-      category,
-      cost: price,
-      payment,
-      content,
-    };
-
-    try {
-      const { status, data } = await createTransaction(
-        accountbookId,
-        newTransaction,
-      );
-      if (status !== 200) throw new Error();
-      addTransaction(data);
-      closeModal();
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
   return (
     <div className="transaction__modal" onClick={onOverLayClicked}>
       <div className="form__container">
         <h2>거래내역 추가</h2>
-        {messageInputVisible ? (
-          <>
-            <button type="button" className="message__input__button">
-              메세지로 추가
-            </button>
-            <form onSubmit={onSubmitClicked}>
-              <DateInput
-                {...{ year, month, today, day, setYear, setMonth, setDay }}
-              />
-              <CategoryInput {...{ categoryPool, setCategory }} />
-              <PaymentInput {...{ paymentPool, setPayment }} />
-              <PriceInput {...{ setPrice }} />
-              <ContentInput {...{ onContentChange }} />
-
-              <div className="transaction__button__container">
-                <button type="submit" className="transaction__submit__button">
-                  확인
-                </button>
-                <input
-                  type="reset"
-                  className="transaction__cancel__button"
-                  onClick={closeModal}
-                  value="취소"
-                />
-              </div>
-            </form>
-          </>
+        {!messageInputVisible ? (
+          <TransactionAddForm
+            {...{
+              setMessageInputVisible,
+              accountbookId,
+              closeModal,
+            }}
+          />
         ) : (
           <>
-            <button type="button" className="message__input__back">
+            <button
+              type="button"
+              className="message__input__back"
+              onClick={() => setMessageInputVisible(false)}
+            >
               뒤로가기
             </button>
-            <MessageInput />
+            <div className="form">
+              <MessageInput {...{ message, setMessage }} />
+              <div className="transaction__button__container">
+                <button type="button" className="transaction__submit__button">
+                  확인
+                </button>
+                <button
+                  type="button"
+                  className="transaction__cancel__button"
+                  onClick={() => setMessage('')}
+                >
+                  지우기
+                </button>
+              </div>
+            </div>
           </>
         )}
       </div>
